[FINAL-1] Tổng hợp và fix code - Bình

+++ app/src/main/java/com/example/cinemaapp/model/Ghe.java
@@ -1,5 +1,7 @@
 package com.example.cinemaapp.model;
 
+import java.text.DecimalFormat;
+
 import ml.huytools.lib.API.ApiProvider;
 import ml.huytools.lib.Annotation.JsonName;
 import ml.huytools.lib.MVP.Model;
@@ -9,7 +11,7 @@ public class Ghe extends Model {
     public static final int STATUS_RESERVED = 2;
     @JsonName
     public String MaGhe;
-    public int Gia;
+    public Double Gia;
     @JsonName
     public int status;
 }
